pyproject.toml
[project]
name = "saliency-gan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "ruff>=0.15.6",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
]
